+++ mirror-soul/src/components/signup/steps/Step2_BasicProfile/hooks/useStep2Form.ts
@@ -1,7 +1,8 @@
 import { useState, useCallback, useRef } from 'react';
 import { Alert } from 'react-native';
 import { Step2State } from '../types/step2';
-import { checkNicknameDuplicate, getPresignedUrl } from '@/src/services/onboardingService';
+import { checkNicknameDuplicate } from '@/src/services/onboardingService';
+import { getPresignedUrl, completeFileUpload } from '@/src/services/fileService';
 import { uploadFileToS3 } from '@/src/services/s3Service';
 import { jobCategories } from '../Professional/jobData';
 
@@ -80,7 +81,17 @@ export function useStep2Form() {
       // 2. S3 직접 업로드
       await uploadFileToS3(presignedUrl, fileUri, contentType);
 
-      // 3. 상태 업데이트
+      // 3. 업로드 완료 알림 (Backend finalize)
+      const completeResponse = await completeFileUpload({
+        objectKey,
+        fileType: 'JOB_CERTIFICATION',
+      });
+
+      if (!completeResponse.isSuccess) {
+        throw new Error(completeResponse.message || '파일 업로드 확인에 실패했습니다.');
+      }
+
+      // 4. 상태 업데이트
       updateState({ 
         isJobVerified: true,
         jobCertificationObjectKey: objectKey
